Split UTs into different files
